+++ steel-core/src/block_entity/entities/end_portal.rs
@@ -1,6 +1,6 @@
 //! End portal block entity.
 
-use std::sync::{Arc, Weak};
+use std::sync::Weak;
 
 use simdnbt::borrow::BaseNbtCompound as BorrowedNbtCompound;
 use simdnbt::owned::NbtCompound;
@@ -8,7 +8,7 @@ use steel_registry::block_entity_type::BlockEntityTypeRef;
 use steel_registry::vanilla_block_entity_types;
 use steel_utils::{BlockPos, BlockStateId, DowncastType, DowncastTypeKey};
 
-use crate::block_entity::BlockEntity;
+use crate::block_entity::{BlockEntity, impl_block_entity_accessors};
 use crate::world::World;
 
 /// Vanilla `TheEndPortalBlockEntity`.
@@ -42,33 +42,7 @@ impl BlockEntity for EndPortalBlockEntity {
         &vanilla_block_entity_types::END_PORTAL
     }
 
-    fn get_block_pos(&self) -> BlockPos {
-        self.pos
-    }
-
-    fn get_block_state(&self) -> BlockStateId {
-        self.state
-    }
-
-    fn set_block_state(&mut self, state: BlockStateId) {
-        self.state = state;
-    }
-
-    fn is_removed(&self) -> bool {
-        self.removed
-    }
-
-    fn set_removed(&mut self) {
-        self.removed = true;
-    }
-
-    fn clear_removed(&mut self) {
-        self.removed = false;
-    }
-
-    fn get_level(&self) -> Option<Arc<World>> {
-        self.level.upgrade()
-    }
+    impl_block_entity_accessors!(level);
 
     fn load_additional(&mut self, _nbt: &BorrowedNbtCompound<'_>) {}
